+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes will be documented in this file.
 
+## v0.0.11
+- new: configurable call and deploy gas
+- new: `.restartblockchain` command to restart ganache e.g. after config changes
+- fix: fixed returnval for some keywords
+- fix: show result for functions declaring multiple return vals
+- fix: naive fix to resolve function declarations for multi returnval function invocations.
+
+<img width="670" alt="image" src="https://user-images.githubusercontent.com/2865694/157049887-d8e8f763-94e4-43fa-bf23-042e2bbe90d9.png">
+
 ## v0.0.10
 - new: update to solc@0.8.11
 - new: basic autocomplete for built-ins (configurable via `.config`) - #11